Avoid the need for python-sh on trigger slaves
Move the function gerrit_project_map into its own module, so it can
avoid the imports from the rest of its old location.
